make readPacket use a bufio.Reader
This should give us some performance wins, since somewhere down the call
chain we read byte-by-byte, which has suboptimal performance when
reading involves a syscall.

Note that even in the recursive case, we are not buffering multiple
times, since the bufio.NewReader just returns the underlying reader if
it is already buffered.